+++ seatable-app/triggers/intern_search_columns.js
@@ -1,24 +1,22 @@
 const searchableColumnTypes = [
-  'text',
-  'long-text',
-  'number',
-  'single-select',
-  'email',
-  'url',
-  'rate',
-  'formula',
+  "text",
+  "long-text",
+  "number",
+  "single-select",
+  "email",
+  "url",
+  "rate",
+  "formula",
 ];
 
 const perform = async (z, bundle) => {
   const returnData = [];
 
   if (!bundle.authData.baseUuid) {
-    console.log("baseUuid is not set or empty...");
     return returnData;
   }
 
   if (!bundle.inputData.table_id) {
-    console.log("table_name is not set or empty...");
     return returnData;
   }
 
@@ -26,13 +24,16 @@ const perform = async (z, bundle) => {
     url: `${bundle.authData.serverUrl}/api-gateway/api/v2/dtables/${bundle.authData.baseUuid}/metadata/`,
   });
 
-  const table = metadata.data.metadata.tables.find(table => table._id === bundle.inputData.table_id);
+  const table = metadata.data.metadata.tables.find(
+    (table) => table._id === bundle.inputData.table_id
+  );
   if (!table) {
     throw new Error(`Table with ID ${bundle.inputData.table_id} not found`);
   }
 
-  const columns = table.columns.filter(column => searchableColumnTypes.includes(column.type))
-    .map(column => ({ id: column.key, name: column.name }));
+  const columns = table.columns
+    .filter((column) => searchableColumnTypes.includes(column.type))
+    .map((column) => ({ id: column.key, name: column.name }));
 
   return columns;
 };